fix init
