+++ molecool/__init__.py
@@ -5,6 +5,10 @@ A Python package for analyzing and visualizing pdb and xyz files. For MolSSI May
 
 # Add imports here
 from .functions import *
+from .molecule import build_bond_list
+from .visualize import  draw_molecule, bond_histogram
+from .measure import calculate_distance, calculate_angle
+from .atom_data import atom_colors, atomic_weights
 
 # Handle versioneer
 from ._version import get_versions
